requirements.txt
openai
openai-whisper
python-dotenv
pvporcupine
pvrhino
numpy
datasets
speechrecognition
transformers
pyaudio
torch
torchvision
torchaudio
opencv-python
pyttsx3
py3-tts
requests
websockets